fix(@chakra-ui/gatsby-plugin): correct fast refresh and theme behavior
