+++ tooling/gatsby-plugin/gatsby-browser.js
@@ -1,31 +1,16 @@
-import React from "react"
-import {
-  ThemeProvider,
-  ColorModeProvider,
-  CSSReset,
-  GlobalStyle,
-  PortalManager,
-} from "@chakra-ui/react"
-import theme from "./src/theme"
+import * as React from "react"
+import { WrapRootElement } from "./src/provider"
 
 export const wrapRootElement = (
   { element },
   { isResettingCSS = true, isUsingColorMode = true, portalZIndex = 40 },
 ) => {
-  const content = (
-    <>
-      {isResettingCSS && <CSSReset />}
-      <GlobalStyle />
-      <PortalManager zIndex={portalZIndex}>{element}</PortalManager>
-    </>
-  )
   return (
-    <ThemeProvider theme={theme}>
-      {isUsingColorMode ? (
-        <ColorModeProvider options={theme.config}>{content}</ColorModeProvider>
-      ) : (
-        content
-      )}
-    </ThemeProvider>
+    <WrapRootElement
+      element={element}
+      isResettingCSS={isResettingCSS}
+      isUsingColorMode={isUsingColorMode}
+      portalZIndex={portalZIndex}
+    />
   )
 }
