+++ website/pages/guides/integrations/with-gatsby.mdx
@@ -69,9 +69,9 @@ export default theme
 Chakra exposes some context providers to give your website context of the theme,
 and color mode.
 
-Create the two files [`gatsby-browser.js`](https://www.gatsbyjs.com/docs/reference/config-files/gatsby-browser/#wrapRootElement) and [`gatsby-ssr.js`](https://www.gatsbyjs.com/docs/reference/config-files/gatsby-ssr/#wrapRootElement) at the root of your Gatsby project.
+Create the two files [`gatsby-browser.js`](https://www.gatsbyjs.com/docs/reference/config-files/gatsby-browser/#wrapRootElement) and [`gatsby-ssr.js`](https://www.gatsbyjs.com/docs/reference/config-files/gatsby-ssr/#wrapRootElement) at the root of your Gatsby project. Also create a `provider.js` that contains a React component that will be used in both files.
 
-Add this to your `gatsby-browser.js`:
+Create the `provider.js` at the root of your project:
 
 ```js live=false
 import React from "react"
@@ -79,7 +79,7 @@ import { ChakraProvider, CSSReset, extendTheme } from "@chakra-ui/react"
 
 const customTheme = extendTheme({});
 
-export const wrapRootElement = ({ element }) => {
+export const WrapRootElement = ({ element }) => {
   return (
     <ChakraProvider theme={customTheme}>
       <CSSReset />
@@ -89,11 +89,26 @@ export const wrapRootElement = ({ element }) => {
 }
 ```
 
+Add this to your `gatsby-browser.js`:
+
+```js live=false
+import React from "react"
+import { WrapRootElement } from "./provider"
+
+export const wrapRootElement = ({ element }) => {
+  return <WrapRootElement element={element} />
+}
+```
+
 And this to your `gatsby-ssr.js`:
 
 ```js live=false
-// gatsby-ssr.js
-export { wrapRootElement } from "./gatsby-browser"
+import React from "react"
+import { WrapRootElement } from "./provider"
+
+export const wrapRootElement = ({ element }) => {
+  return <WrapRootElement element={element} />
+}
 ```
 
 ### v0.8 and below
@@ -106,7 +121,7 @@ import React from "react"
 import { ChakraProvider, CSSReset } from "@chakra-ui/react"
 import customTheme from "./theme/theme"
 
-export const wrapRootElement = ({ element }) => {
+export const WrapRootElement = ({ element }) => {
   return (
     <>
       <ThemeProvider theme={customTheme}>
